Add span IDs that uniquely identify spans
This branch introduces a `span::Id` type and changes span equality to be
based on the span's `Id`. The `Id` type is used to identify the span
within the context of the process.

These span IDs are intended to be used as a source of truth for
determining whether two span handles are equal. Previously, this
was done based on the span's creation timestamp and metadata, but the
creation timestamps are slated for removal (see #24). When they are
removed, the metadata alone will not be sufficient to uniquely identify
the span, since that would mean two distinct spans created at the same
call site would be "equal".

Similarly, we cannot base span equality on their field values, for a
number of reasons:
+ Two spans with the same metadata and no fields would be equal,
  even if they are distinct periods of time.
+ Similarly, two spans which happen to have the same field values, but
  represent distinct periods of time would also be incorrectly equal.
+ Using field values for equality and hashing would require that
 `Value`s implement `Hash` and `PartialEq`, which they cannot do when
  they are represented by trait objects (as these traits are not object
  safe).
+ Also, comparing the field values would require iterating across all
  the fields.

Currently, the `Id` type is represented by a `u64`. However, the `Id`
type is opaque to downstream users, beyond the contract of implementing
`Clone`, `Hash`, and `PartialEq`, and being able to be _constructed_
from a `u64`, so we could swap out this implementation with something
else later fairly transparently, as long as we can construct it from a
`u64`.

`Subscriber`s are responsible for generating the IDs for each span
(through the `Subscriber::new_span_id` trait method), allowing them to
choose the span ID generation strategy that supports their use case.

Required for #24