Change `Dispatch::current` to select the current _span_'s dispatcher
If avalable, the current span's dispatcher will be used. Otherwise, the
thread's default dispatcher is used. This allows subscribers to collect whole trace
trees, even across threads.

Also, this adds a `Default` impl to `Dispatch` that always returns the thread's
default dispatcher. This should be used outside of a trace tree.